% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanLyrics.R
\name{cleanLyrics}
\alias{cleanLyrics}
\title{cleanLyrics}
\usage{
cleanLyrics(lyr, expContractions = FALSE)
}
\arguments{
\item{lyr}{data frame of formatted lyrics from formatLyrics()}

\item{expContractions}{boolean value that indicates if the user wants to expand the contractions}
}
\value{
data frame of cleaned lyrics
}
\description{
Expands contractions and removes punctuation from the scraped and formatted lyrics
}
\examples{
\dontrun{
cleaned_lyrics <- cleanLyrics(formatted_lyrics)
}
}
