% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatLyrics.R
\name{formatLyrics}
\alias{formatLyrics}
\title{formatLyrics}
\usage{
formatLyrics(lyrics)
}
\arguments{
\item{lyrics}{matrix/array of the scraped lyrics from getSong() or getArtist()}
}
\value{
data frame of formatted lyrics
}
\description{
Formats the scraped lyrics into a data frame and removes new line characters
}
\examples{
\dontrun{
formatted_lyrics <- formatLyrics(billieLyrics)
}

}
