% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dtm.R
\name{get_dtm}
\alias{get_dtm}
\title{get_dtm}
\usage{
get_dtm(cleanedlyrics)
}
\arguments{
\item{cleanedlyrics}{data frame of formatted and cleaned lyrics by formatLyrics() and cleanLyrics()}
}
\value{
Document Term Matrix of lyrics
}
\description{
Transforms the data frame of formatted and cleaned lyrics into a Document Term Matrix
}
\examples{
\dontrun{
my_dtm <- get_dtm(cleaned_lyrics)
}

}
