% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wordcloud.R
\name{get_wordcloud}
\alias{get_wordcloud}
\title{get_wordcloud}
\usage{
get_wordcloud(cleanedlyrics = NULL, dtm = FALSE)
}
\arguments{
\item{cleanedlyrics}{data frame of formatted and cleaned lyrics by formatLyrics() and cleanLyrics() or a Document Term Matrix of lyrics}

\item{dtm}{boolean that indicates if cleanedlyrics is a data frame or Document Term Matrix}
}
\value{
plot, word cloud of lyrics
}
\description{
Creates a wordcloud from a Document Term Matrix or data frame of formatted and cleaned lyrics
}
\examples{
\dontrun{
get_wordcloud(cleaned_lyrics, dtm = F)
get_wordcloud(my_dtm, dtm = T)
}

}
