% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtist.R
\name{getArtist}
\alias{getArtist}
\title{getArtist}
\usage{
getArtist(artist)
}
\arguments{
\item{artist}{Name of the desired artist}
}
\value{
matrix/array of scraped lyrics
}
\description{
Retrieves all the songs from a specific artist
}
\examples{
billieLyrics <- getArtist("Billie Eilish")

}
