% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSong.R
\name{getSong}
\alias{getSong}
\title{getSong}
\usage{
getSong(name, artist)
}
\arguments{
\item{name}{Name of the desired song}

\item{artist}{Name of the desired artist}
}
\value{
matrix/array of scraped lyrics
}
\description{
Retrieves a specific song from a specific artist
}
\examples{
song <- getSong('Billie Eilish', 'Party Favor')

}
