% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentProportion.R
\name{sentimentProportion}
\alias{sentimentProportion}
\title{sentimentProportion}
\usage{
sentimentProportion(x)
}
\arguments{
\item{x}{vector, or a column of a data frame, of lyrics}
}
\description{
Calculates the detected sentiments in the lyrics
}
\examples{
\dontrun{
sentimentProportion(cleaned_lyrics[,1])
}

}
