% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5words.R
\name{top5words}
\alias{top5words}
\title{top5words}
\usage{
top5words(cleanedlyrics, dtm = FALSE)
}
\arguments{
\item{cleanedlyrics}{data frame of formatted and cleaned lyrics by formatLyrics() and cleanLyrics() or a Document Term Matrix of lyrics}

\item{dtm}{boolean that indicates if cleanedlyrics is a data frame or Document Term Matrix}
}
\value{
bar plot of top five most used lyrics in data frame
}
\description{
Creates a bar plot of the top 5 most used words in a Document Term Matrix or data frame of formatted and cleaned lyrics
}
\examples{
\dontrun{
top5words(cleaned_lyrics, dtm = F)
top5words(my_dtm, dtm = T)
}

}
